refactor(worker): extract agent-turn finalization
Move the attempt teardown finally body into finalizeTurnAttempt so the
quiescence, lease-release, capture, and sandbox-release order stays in
one dedicated module.
